feat: introduce raw engine
