+++ packages/engine-javascript/build.config.ts
@@ -3,6 +3,8 @@ import { defineBuildConfig } from 'unbuild'
 export default defineBuildConfig({
   entries: [
     'src/index.ts',
+    'src/engine-compile.ts',
+    'src/engine-raw.ts',
   ],
   declaration: true,
   rollup: {
