+++ packages/engine-javascript/package.json
@@ -21,6 +21,10 @@
     ".": {
       "types": "./dist/index.d.mts",
       "default": "./dist/index.mjs"
+    },
+    "./raw": {
+      "types": "./dist/raw.d.mts",
+      "default": "./dist/raw.mjs"
     }
   },
   "main": "./dist/index.mjs",
