+++ packages/langs-precompiled/scripts/precompile.ts
@@ -1,6 +1,6 @@
 import type { LanguageRegistration } from '@shikijs/types'
 import { traverseGrammarPatterns } from '../../../scripts/utils'
-import { defaultJavaScriptRegexConstructor } from '../../engine-javascript/src/index'
+import { defaultJavaScriptRegexConstructor } from '../../engine-javascript/src/engine-compile'
 
 export function precompileGrammar(grammar: LanguageRegistration): LanguageRegistration {
   const precompiled: LanguageRegistration = structuredClone(grammar)
