+++ packages/shiki/test/core-sync.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it } from 'vitest'
-import { createJavaScriptRegexEngine } from '../../engine-javascript/src'
+import { createJavaScriptRegexEngine } from '../../engine-javascript/src/engine-compile'
 import { createHighlighterCoreSync } from '../src/core'
 import js from '../src/langs/javascript.mjs'
 import nord from '../src/themes/nord.mjs'
